ui: Move commands from uiMain to the CoreCommands plugin
