+++ ui/src/frontend/ui_main.ts
@@ -13,47 +13,33 @@
 // limitations under the License.
 
 import m from 'mithril';
-import {copyToClipboard} from '../base/clipboard';
+import {DisposableStack} from '../base/disposable_stack';
 import {findRef} from '../base/dom_utils';
 import {FuzzyFinder} from '../base/fuzzy';
 import {assertExists, assertUnreachable} from '../base/logging';
+import {Icons} from '../base/semantic_icons';
 import {undoCommonChatAppReplacements} from '../base/string_utils';
+import {addQueryResultsTab} from '../components/query_table/query_result_tab';
+import {AppImpl} from '../core/app_impl';
+import {CookieConsent} from '../core/cookie_consent';
+import {featureFlags} from '../core/feature_flags';
+import {OmniboxMode} from '../core/omnibox_manager';
+import {TraceImpl} from '../core/trace_impl';
 import {Command} from '../public/command';
+import {Anchor} from '../widgets/anchor';
+import {Button} from '../widgets/button';
 import {HotkeyConfig, HotkeyContext} from '../widgets/hotkey_context';
 import {HotkeyGlyphs} from '../widgets/hotkey_glyphs';
+import {LinearProgress} from '../widgets/linear_progress';
 import {maybeRenderFullscreenModalDialog, showModal} from '../widgets/modal';
-import {CookieConsent} from '../core/cookie_consent';
+import {Spinner} from '../widgets/spinner';
+import {initCssConstants} from './css_constants';
 import {toggleHelp} from './help_modal';
 import {Omnibox, OmniboxOption} from './omnibox';
-import {addQueryResultsTab} from '../components/query_table/query_result_tab';
 import {Sidebar} from './sidebar';
-import {Topbar} from './topbar';
-import {shareTrace} from './trace_share_utils';
-import {OmniboxMode} from '../core/omnibox_manager';
-import {DisposableStack} from '../base/disposable_stack';
-import {Spinner} from '../widgets/spinner';
-import {TraceImpl} from '../core/trace_impl';
-import {AppImpl} from '../core/app_impl';
-import {getTimeSpanOfSelectionOrVisibleWindow} from '../public/utils';
-import {DurationPrecision, TimestampFormat} from '../public/timeline';
-import {Workspace} from '../public/workspace';
-import {
-  deserializeAppStatePhase1,
-  deserializeAppStatePhase2,
-  JsonSerialize,
-  parseAppState,
-  serializeAppState,
-} from '../core/state_serialization';
-import {featureFlags} from '../core/feature_flags';
-import {trackMatchesFilter} from '../core/track_manager';
 import {renderStatusBar} from './statusbar';
-import {formatTimezone, timezoneOffsetMap} from '../base/time';
-import {LinearProgress} from '../widgets/linear_progress';
 import {taskTracker} from './task_tracker';
-import {Button} from '../widgets/button';
-import {initCssConstants} from './css_constants';
-import {Anchor} from '../widgets/anchor';
-import {Icons} from '../base/semantic_icons';
+import {Topbar} from './topbar';
 
 const showStatusBarFlag = featureFlags.register({
   id: 'Enable status bar',
@@ -61,7 +47,6 @@ const showStatusBarFlag = featureFlags.register({
   defaultValue: true,
 });
 
-const QUICKSAVE_LOCALSTORAGE_KEY = 'quicksave';
 const OMNIBOX_INPUT_REF = 'omnibox';
 
 // This wrapper creates a new instance of UiMainPerTrace for each new trace
@@ -119,410 +104,6 @@ export class UiMainPerTrace implements m.ClassComponent {
     if (trace === undefined) return;
     document.title = `${trace.traceInfo.traceTitle || 'Trace'} - Perfetto UI`;
     this.maybeShowJsonWarning();
-
-    const cmds: Command[] = [
-      {
-        id: 'dev.perfetto.SetTimestampFormat',
-        name: 'Set timestamp and duration format',
-        callback: async () => {
-          const TF = TimestampFormat;
-          const timeZone = formatTimezone(trace.traceInfo.tzOffMin);
-          const result = await app.omnibox.prompt('Select format...', {
-            values: [
-              {format: TF.Timecode, name: 'Timecode'},
-              {format: TF.UTC, name: 'Realtime (UTC)'},
-
-              {format: TF.TraceTz, name: `Realtime (Trace TZ - ${timeZone})`},
-              {format: TF.Seconds, name: 'Seconds'},
-              {format: TF.Milliseconds, name: 'Milliseconds'},
-              {format: TF.Microseconds, name: 'Microseconds'},
-              {format: TF.TraceNs, name: 'Trace nanoseconds'},
-              {
-                format: TF.TraceNsLocale,
-                name: 'Trace nanoseconds (with locale-specific formatting)',
-              },
-              {format: TF.CustomTimezone, name: 'Custom Timezone'},
-            ],
-            getName: (x) => x.name,
-          });
-          if (!result) return;
-
-          if (result.format === TF.CustomTimezone) {
-            const result = await app.omnibox.prompt('Select format...', {
-              values: Object.entries(timezoneOffsetMap),
-              getName: ([key]) => key,
-            });
-
-            if (!result) return;
-            trace.timeline.timezoneOverride.set(result[0]);
-          }
-
-          trace.timeline.timestampFormat = result.format;
-        },
-      },
-      {
-        id: 'dev.perfetto.SetDurationPrecision',
-        name: 'Set duration precision',
-        callback: async () => {
-          const DF = DurationPrecision;
-          const result = await app.omnibox.prompt(
-            'Select duration precision mode...',
-            {
-              values: [
-                {format: DF.Full, name: 'Full'},
-                {format: DF.HumanReadable, name: 'Human readable'},
-              ],
-              getName: (x) => x.name,
-            },
-          );
-          result && (trace.timeline.durationPrecision = result.format);
-        },
-      },
-      {
-        id: 'dev.perfetto.TogglePerformanceMetrics',
-        name: 'Toggle performance metrics',
-        callback: () =>
-          (app.perfDebugging.enabled = !app.perfDebugging.enabled),
-      },
-      {
-        id: 'dev.perfetto.ShareTrace',
-        name: 'Share trace',
-        callback: () => shareTrace(trace),
-      },
-      {
-        id: 'dev.perfetto.SearchNext',
-        name: 'Go to next search result',
-        callback: () => {
-          trace.search.stepForward();
-        },
-        defaultHotkey: 'Enter',
-      },
-      {
-        id: 'dev.perfetto.SearchPrev',
-        name: 'Go to previous search result',
-        callback: () => {
-          trace.search.stepBackwards();
-        },
-        defaultHotkey: 'Shift+Enter',
-      },
-      {
-        id: 'dev.perfetto.SwitchToQueryMode',
-        name: 'Switch to query mode',
-        callback: () => trace.omnibox.setMode(OmniboxMode.Query),
-      },
-      {
-        id: 'dev.perfetto.RunQuery',
-        name: 'Runs an SQL query',
-        callback: async (rawSql: unknown) => {
-          const query =
-            typeof rawSql === 'string'
-              ? rawSql
-              : await trace.omnibox.prompt('Enter SQL...');
-          if (!query) {
-            return;
-          }
-          await trace.engine.query(query);
-        },
-      },
-      {
-        id: 'dev.perfetto.RunQueryAndShowTab',
-        name: 'Runs an SQL query and opens results in a tab',
-        callback: async (rawSql: unknown) => {
-          const query =
-            typeof rawSql === 'string'
-              ? rawSql
-              : await trace.omnibox.prompt('Enter SQL...');
-          if (!query) {
-            return;
-          }
-          addQueryResultsTab(trace, {
-            query,
-            title: 'Command Query',
-          });
-        },
-      },
-      {
-        id: 'dev.perfetto.SwitchToSearchMode',
-        name: 'Switch to search mode',
-        callback: () => trace.omnibox.setMode(OmniboxMode.Search),
-        defaultHotkey: '/',
-      },
-      {
-        id: 'dev.perfetto.CopyTimeWindow',
-        name: `Copy selected time window to clipboard`,
-        callback: async () => {
-          const window = await getTimeSpanOfSelectionOrVisibleWindow(trace);
-          const query = `ts >= ${window.start} and ts < ${window.end}`;
-          copyToClipboard(query);
-        },
-      },
-      {
-        id: 'dev.perfetto.FocusSelection',
-        name: 'Focus current selection',
-        callback: () => trace.selection.scrollToSelection(),
-        defaultHotkey: 'F',
-      },
-      {
-        id: 'dev.perfetto.ZoomOnSelection',
-        name: 'Zoom in on current selection',
-        callback: () => trace.selection.zoomOnSelection(),
-      },
-      {
-        id: 'dev.perfetto.Deselect',
-        name: 'Deselect',
-        callback: () => {
-          trace.selection.clearSelection();
-        },
-        defaultHotkey: 'Escape',
-      },
-      {
-        id: 'dev.perfetto.NextFlow',
-        name: 'Next flow',
-        callback: () => trace.flows.focusOtherFlow('Forward'),
-        defaultHotkey: 'Mod+]',
-      },
-      {
-        id: 'dev.perfetto.PrevFlow',
-        name: 'Prev flow',
-        callback: () => trace.flows.focusOtherFlow('Backward'),
-        defaultHotkey: 'Mod+[',
-      },
-      {
-        id: 'dev.perfetto.MoveNextFlow',
-        name: 'Move next flow',
-        callback: () => trace.flows.moveByFocusedFlow('Forward'),
-        defaultHotkey: ']',
-      },
-      {
-        id: 'dev.perfetto.MovePrevFlow',
-        name: 'Move prev flow',
-        callback: () => trace.flows.moveByFocusedFlow('Backward'),
-        defaultHotkey: '[',
-      },
-
-      // Provides a test bed for resolving events using a SQL table name and ID
-      // which is used in deep-linking, amongst other places.
-      {
-        id: 'dev.perfetto.SelectEventByTableNameAndId',
-        name: 'Select event by table name and ID',
-        callback: async () => {
-          const rootTableName = await trace.omnibox.prompt('Enter table name');
-          if (rootTableName === undefined) return;
-
-          const id = await trace.omnibox.prompt('Enter ID');
-          if (id === undefined) return;
-
-          const num = Number(id);
-          if (!isFinite(num)) return; // Rules out NaN or +-Infinity
-
-          trace.selection.selectSqlEvent(rootTableName, num, {
-            scrollToSelection: true,
-          });
-        },
-      },
-      {
-        id: 'dev.perfetto.SelectAll',
-        name: 'Select all',
-        callback: () => {
-          // This is a dual state command:
-          // - If one ore more tracks are already area selected, expand the time
-          //   range to include the entire trace, but keep the selection on just
-          //   these tracks.
-          // - If nothing is selected, or all selected tracks are entirely
-          //   selected, then select the entire trace. This allows double tapping
-          //   Ctrl+A to select the entire track, then select the entire trace.
-          let tracksToSelect: ReadonlyArray<string>;
-          const selection = trace.selection.selection;
-          if (selection.kind === 'area') {
-            // Something is already selected, let's see if it covers the entire
-            // span of the trace or not
-            const coversEntireTimeRange =
-              trace.traceInfo.start === selection.start &&
-              trace.traceInfo.end === selection.end;
-            if (!coversEntireTimeRange) {
-              // If the current selection is an area which does not cover the
-              // entire time range, preserve the list of selected tracks and
-              // expand the time range.
-              tracksToSelect = selection.trackUris;
-            } else {
-              // If the entire time range is already covered, update the selection
-              // to cover all tracks.
-              tracksToSelect = trace.workspace.flatTracks
-                .map((t) => t.uri)
-                .filter((uri) => uri !== undefined);
-            }
-          } else {
-            // If the current selection is not an area, select all.
-            tracksToSelect = trace.workspace.flatTracks
-              .map((t) => t.uri)
-              .filter((uri) => uri !== undefined);
-          }
-          const {start, end} = trace.traceInfo;
-          trace.selection.selectArea({
-            start,
-            end,
-            trackUris: tracksToSelect,
-          });
-        },
-        defaultHotkey: 'Mod+A',
-      },
-      {
-        id: 'dev.perfetto.ConvertSelectionToArea',
-        name: 'Convert selection to area selection',
-        callback: () => {
-          const selection = trace.selection.selection;
-          const range = trace.selection.getTimeSpanOfSelection();
-          if (selection.kind === 'track_event' && range) {
-            trace.selection.selectArea({
-              start: range.start,
-              end: range.end,
-              trackUris: [selection.trackUri],
-            });
-          }
-        },
-        defaultHotkey: 'R',
-      },
-      {
-        id: 'dev.perfetto.ToggleDrawer',
-        name: 'Toggle drawer',
-        defaultHotkey: 'Q',
-        callback: () => trace.tabs.toggleTabPanelVisibility(),
-      },
-      {
-        id: 'dev.perfetto.CopyPinnedToWorkspace',
-        name: 'Copy pinned tracks to workspace',
-        callback: async () => {
-          const pinnedTracks = trace.workspace.pinnedTracks;
-          if (!pinnedTracks.length) {
-            window.alert('No pinned tracks to copy');
-            return;
-          }
-
-          const ws = await this.selectWorkspace(trace, 'Pinned tracks');
-          if (!ws) return;
-
-          for (const pinnedTrack of pinnedTracks) {
-            const clone = pinnedTrack.clone();
-            ws.addChildLast(clone);
-          }
-          trace.workspaces.switchWorkspace(ws);
-        },
-      },
-      {
-        id: 'dev.perfetto.CopyFilteredToWorkspace',
-        name: 'Copy filtered tracks to workspace',
-        callback: async () => {
-          // Copies all filtered tracks as a flat list to a new workspace. This
-          // means parents are not included.
-          const tracks = trace.workspace.flatTracks.filter((track) =>
-            trackMatchesFilter(trace, track),
-          );
-
-          if (!tracks.length) {
-            window.alert('No filtered tracks to copy');
-            return;
-          }
-
-          const ws = await this.selectWorkspace(trace, 'Filtered tracks');
-          if (!ws) return;
-
-          for (const track of tracks) {
-            const clone = track.clone();
-            ws.addChildLast(clone);
-          }
-          trace.workspaces.switchWorkspace(ws);
-        },
-      },
-      {
-        id: 'dev.perfetto.CopySelectedTracksToWorkspace',
-        name: 'Copy selected tracks to workspace',
-        callback: async () => {
-          const selection = trace.selection.selection;
-
-          if (selection.kind !== 'area' || selection.trackUris.length === 0) {
-            window.alert('No selected tracks to copy');
-            return;
-          }
-
-          const workspace = await this.selectWorkspace(trace);
-          if (!workspace) return;
-
-          for (const uri of selection.trackUris) {
-            const node = trace.workspace.getTrackByUri(uri);
-            if (!node) continue;
-            const newNode = node.clone();
-            workspace.addChildLast(newNode);
-          }
-          trace.workspaces.switchWorkspace(workspace);
-        },
-      },
-      {
-        id: 'dev.perfetto.Quicksave',
-        name: 'Quicksave UI state to localStorage',
-        callback: () => {
-          const state = serializeAppState(trace);
-          const json = JsonSerialize(state);
-          localStorage.setItem(QUICKSAVE_LOCALSTORAGE_KEY, json);
-        },
-      },
-      {
-        id: 'dev.perfetto.Quickload',
-        name: 'Quickload UI state from the localStorage',
-        callback: () => {
-          const json = localStorage.getItem(QUICKSAVE_LOCALSTORAGE_KEY);
-          if (json === null) {
-            showModal({
-              title: 'Nothing saved in the quicksave slot',
-              buttons: [{text: 'Dismiss'}],
-            });
-            return;
-          }
-          const parsed = JSON.parse(json);
-          const state = parseAppState(parsed);
-          if (state.ok) {
-            deserializeAppStatePhase1(state.value, trace);
-            deserializeAppStatePhase2(state.value, trace);
-          }
-        },
-      },
-      {
-        id: `${app.pluginId}#RestoreDefaults`,
-        name: 'Reset all flags back to default values',
-        callback: () => {
-          featureFlags.resetAll();
-          window.location.reload();
-        },
-      },
-    ];
-
-    // Register each command with the command manager
-    cmds.forEach((cmd) => {
-      this.trash.use(trace.commands.registerCommand(cmd));
-    });
-  }
-
-  // Selects a workspace or creates a new one.
-  private async selectWorkspace(
-    trace: TraceImpl,
-    newWorkspaceName = 'Untitled workspace',
-  ): Promise<Workspace | undefined> {
-    const options = trace.workspaces.all
-      .filter((ws) => ws.userEditable)
-      .map((ws) => ({title: ws.title, fn: () => ws}))
-      .concat([
-        {
-          title: 'New workspace...',
-          fn: () => trace.workspaces.createEmptyWorkspace(newWorkspaceName),
-        },
-      ]);
-
-    const result = await trace.omnibox.prompt('Select a workspace...', {
-      values: options,
-      getName: (ws) => ws.title,
-    });
-
-    if (!result) return undefined;
-    return result.fn();
   }
 
   private renderOmnibox(): m.Children {
